Moved gallery-specific bits of UserExtensions back to gallery
